\name{MillerCalib}
\alias{MillerCalib}
\title{Miller's model calibration satistics}
\description{
This function is DEPRECATED, please use \code{\link{Miller}} instead.
}
\usage{
MillerCalib(model = NULL, obs = NULL, pred = NULL, plot = TRUE,
line.col = "darkblue", diag = TRUE, diag.col = "lightblue3", 
plot.values = TRUE, digits = 2, xlab = "", ylab = "", 
main = "Miller calibration", na.rm = TRUE, rm.dup = FALSE, verbosity = 2, ...)
}
\arguments{
  \item{model}{see \code{\link{Miller}}}
  \item{obs}{see \code{\link{Miller}}}
  \item{pred}{see \code{\link{Miller}}}
  \item{plot}{see \code{\link{Miller}}}
  \item{line.col}{see \code{\link{Miller}}}
  \item{diag}{see \code{\link{Miller}}}
  \item{diag.col}{see \code{\link{Miller}}}
  \item{plot.values}{see \code{\link{Miller}}}
  \item{digits}{see \code{\link{Miller}}}
  \item{xlab}{see \code{\link{Miller}}}
  \item{ylab}{see \code{\link{Miller}}}
  \item{main}{see \code{\link{Miller}}}
  \item{na.rm}{see \code{\link{Miller}}}
  \item{rm.dup}{see \code{\link{Miller}}}
  \item{verbosity}{see \code{\link{Miller}}}
  \item{\dots}{see \code{\link{Miller}}}
}
\details{see \code{\link{Miller}}}
\value{see \code{\link{Miller}}}
\references{see \code{\link{Miller}}}
\author{A. Marcia Barbosa}
\seealso{
\code{\link{Miller}}
}
\examples{
# This function is DEPRECATED, please see ?Miller instead.
}
